(wip) std mode
