+++ tool/dsqss/__init__.py
@@ -1,11 +1,4 @@
-# from dsqss.read_keyvalues import *
-# from dsqss.hamiltonian import generate_hamiltonian
-# from dsqss.lattice import generate_lattice
-# import dsqss.lattice
-
 import dsqss
-import dsqss.hamgen
-import dsqss.latgen
 import dsqss.hamiltonian
 import dsqss.lattice
