+++ tool/dsqss/std_lattice.py
@@ -1,0 +1,9 @@
+from dsqss.util import ERROR
+from dsqss.hypercubic import HyperCubicLattice
+
+def std_lattice(param):
+    if param['Lattice'] == 'hypercubic':
+        return HyperCubicLattice(param)
+    else:
+        ERROR('Unknown lattice: param["lattice"] = {0}'.format(param['lattice'].lower()))
+
